requirements.txt
Django
djangorestframework
django-filter
path.py
requests